requirements.txt
python-dotenv
streamlit
langchain
langgraph
langchain-core
langchain-community
langchain-openai
langchain_tavily
duckduckgo-search